HW 3/flight, booking/flight-service/requirements.txt
grpcio==1.62.0
grpcio-tools==1.62.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1
